Revert "Move `createSession` and `readSession` to session stores"
