+++ packages/acceptance-tests/src/additional/typeorm.mongodb-store.spec.ts
@@ -6,6 +6,7 @@ import {
   Config,
   Context,
   createApp,
+  createSession,
   dependency,
   Get,
   hashPassword,
@@ -119,7 +120,7 @@ describe('[Sample] TypeORM & MongoDB Store', async () => {
         return new HttpResponseUnauthorized();
       }
 
-      const session = await this.store.createSession();
+      const session = await createSession(this.store);
       session.setUser(user);
       await session.commit();
