+++ packages/acceptance-tests/src/docs/authentication-and-access-control/quick-start/authenticating-users-in-a-stateful-spa-using-header.feature.ts
@@ -11,6 +11,7 @@ import {
   Context,
   controller,
   createApp,
+  createSession,
   dependency,
   Get,
   hashPassword,
@@ -70,7 +71,7 @@ describe('Feature: Authenticating users in a stateful SPA using the `Authorizati
       user.password = await hashPassword(ctx.request.body.password);
       await user.save();
 
-      ctx.session = await this.store.createSession();
+      ctx.session = await createSession(this.store);
       ctx.session.setUser(user);
 
       return new HttpResponseOK({
@@ -91,7 +92,7 @@ describe('Feature: Authenticating users in a stateful SPA using the `Authorizati
         return new HttpResponseUnauthorized();
       }
 
-      ctx.session = await this.store.createSession();
+      ctx.session = await createSession(this.store);
       ctx.session.setUser(user);
 
       return new HttpResponseOK({
