+++ packages/acceptance-tests/src/docs/authentication-and-access-control/session-tokens/destroying-the-session.feature.ts
@@ -12,9 +12,11 @@ import {
   Context,
   controller,
   createApp,
+  createSession,
   HttpResponseNoContent,
   IAppController,
   Post,
+  readSession,
   ServiceManager,
   Store,
   UseSessions
@@ -67,10 +69,10 @@ describe('Feature: Destroying the session', () => {
     dataSource = await createAndInitializeDataSource([ DatabaseSession ]);
     const store = services.get(Store);
 
-    const session = await store.createSession();
+    const session = await createSession(store);
     await session.commit();
 
-    notStrictEqual(await store.readSession(session.getToken()), null);
+    notStrictEqual(await readSession(store, session.getToken()), null);
 
     await request(app)
       .post('/logout')
@@ -83,7 +85,7 @@ describe('Feature: Destroying the session', () => {
       .send()
       .expect(204);
 
-    strictEqual(await store.readSession(session.getToken()), null);
+    strictEqual(await readSession(store, session.getToken()), null);
   });
 
 });
