+++ packages/acceptance-tests/src/docs/authentication-and-access-control/session-tokens/reading-a-session-from-a-token.feature.ts
@@ -8,6 +8,8 @@ import { DataSource } from 'typeorm';
 import {
   Config,
   createService,
+  createSession,
+  readSession,
   Store,
 } from '@foal/core';
 import { DatabaseSession } from '@foal/typeorm';
@@ -34,7 +36,7 @@ describe('Feature: Reading a session from a token', () => {
     async function getFoo(token: string): Promise<any> {
       /* ======================= DOCUMENTATION BEGIN ======================= */
 
-      const session = await store.readSession(token);
+      const session = await readSession(store, token);
       if (!session) {
         throw new Error('Session does not exist or has expired.');
       }
@@ -47,7 +49,7 @@ describe('Feature: Reading a session from a token', () => {
 
     dataSource = await createAndInitializeDataSource([ DatabaseSession ]);
 
-    const session = await store.createSession();
+    const session = await createSession(store);
     session.set('foo', 'bar');
     await session.commit();
