+++ packages/acceptance-tests/src/docs/authentication-and-access-control/session-tokens/regenerating-the-session-id.feature.ts
@@ -10,6 +10,7 @@ import {
   Config,
   Context,
   createApp,
+  createSession,
   dependency,
   Get,
   HttpResponseOK,
@@ -44,7 +45,7 @@ describe('Feature: Regenerating the session ID', () => {
 
       @Get('/new-session')
       async createSessionAndReturnToken(ctx: Context) {
-        ctx.session = await this.store.createSession();
+        ctx.session = await createSession(this.store);
         return new HttpResponseOK({ token: ctx.session.getToken() });
       }
