+++ packages/acceptance-tests/src/docs/authentication-and-access-control/session-tokens/using-the-authorization-header.feature.ts
@@ -11,6 +11,7 @@ import {
   Context,
   controller,
   createApp,
+  createSession,
   dependency,
   Get,
   HttpResponseOK,
@@ -54,7 +55,7 @@ describe('Feature: Using the Authorization header', () => {
       async login(ctx: Context) {
         // Check the user credentials...
 
-        ctx.session = await this.store.createSession();
+        ctx.session = await createSession(this.store);
 
         // See the "authentication" section below
         // to see how to associate a user to the session.
@@ -134,7 +135,7 @@ describe('Feature: Using the Authorization header', () => {
       async login(ctx: Context) {
         // Check the user credentials...
 
-        ctx.session = await this.store.createSession();
+        ctx.session = await createSession(this.store);
 
         // See the "authentication" section below
         // to see how to associate a user to the session.
