+++ packages/acceptance-tests/src/docs/authentication-and-access-control/social-auth/using-social-auth-with-sessions.feature.ts
@@ -9,10 +9,10 @@ import {
   Config,
   Context,
   createController,
+  createSession,
   dependency,
   Get,
   HttpResponseRedirect,
-  SessionState,
   Store,
   UseSessions,
 } from '@foal/core';
@@ -23,27 +23,6 @@ import { DatabaseSession } from '@foal/typeorm';
 describe('Feature: Using social auth with sessions', () => {
 
   let dataSource: DataSource;
-
-  class ConcreteStore extends Store {
-    save(state: SessionState, maxInactivity: number): Promise<void> {
-      throw new Error('Method not implemented.');
-    }
-    read(id: string): Promise<SessionState | null> {
-      throw new Error('Method not implemented.');
-    }
-    update(state: SessionState, maxInactivity: number): Promise<void> {
-      throw new Error('Method not implemented.');
-    }
-    destroy(id: string): Promise<void> {
-      throw new Error('Method not implemented.');
-    }
-    clear(): Promise<void> {
-      throw new Error('Method not implemented.');
-    }
-    cleanUpExpiredSessions(maxInactivity: number, maxLifeTime: number): Promise<void> {
-      throw new Error('Method not implemented.');
-    }
-  }
 
   before(() => {
     Config.set('settings.session.store', getTypeORMStorePath());
@@ -142,7 +121,7 @@ describe('Feature: Using social auth with sessions', () => {
         code: 'known_user'
       }
     });
-    ctx.session = await new ConcreteStore().createSession();
+    ctx.session = await createSession({} as any);
 
     strictEqual(ctx.session.userId, null);
 
@@ -156,7 +135,7 @@ describe('Feature: Using social auth with sessions', () => {
         code: 'unknown_user'
       }
     });
-    ctx2.session = await new ConcreteStore().createSession();
+    ctx2.session = await createSession({} as any);
 
     strictEqual(ctx2.session.userId, null);
     strictEqual(await User.findOneBy({ email: 'unknown@foalts.org' }), null);
