+++ packages/acceptance-tests/src/docs/security/csrf/spa.stateful.spec.ts
@@ -11,6 +11,7 @@ import {
   Context,
   controller,
   createApp,
+  createSession,
   dependency,
   HttpResponseCreated,
   HttpResponseNoContent,
@@ -52,7 +53,7 @@ describe('Feature: Stateful CSRF protection in a Single-Page Application', () =>
         return new HttpResponseUnauthorized();
       }
 
-      ctx.session = ctx.session || await this.store.createSession();
+      ctx.session = ctx.session || await createSession(this.store);
       ctx.session.setUser(user);
 
       return new HttpResponseNoContent();
