+++ packages/core/src/index.ts
@@ -224,4 +224,6 @@ export {
   SessionStore,
   Store,
   UseSessions,
+  createSession,
+  readSession,
 } from './sessions';
