+++ packages/core/src/sessions/core/index.ts
@@ -1,4 +1,6 @@
 export { SESSION_DEFAULT_ABSOLUTE_TIMEOUT, SESSION_DEFAULT_INACTIVITY_TIMEOUT } from './constants';
+export { createSession } from './create-session';
+export { readSession } from './read-session';
 export { Store, SessionStore, SessionAlreadyExists } from './session-store';
 export { Session } from './session';
 export { SessionState } from './session-state.interface';
