+++ packages/core/src/sessions/http/check-user-id-type.ts
