+++ packages/core/src/sessions/http/use-sessions.hook.spec.ts
@@ -30,6 +30,7 @@ import {
 import {
   SESSION_DEFAULT_ABSOLUTE_TIMEOUT,
   SESSION_DEFAULT_INACTIVITY_TIMEOUT,
+  readSession,
   Session,
   SessionState,
   SessionStore,
@@ -512,10 +513,10 @@ describe('UseSessions', () => {
       it('should throw an error if the session state has no CSRF token.', async () => {
         ctx = createContextWithPostMethod({}, { [SESSION_DEFAULT_COOKIE_NAME]: anonymousSessionID });
         return rejects(
-          async () => hook(ctx, services),
+          () => hook(ctx, services),
           {
             message: 'Unexpected error: the session content does not have a "csrfToken" field. '
-              + 'Are you sure you created the session with "Store.createSession"?'
+              + 'Are you sure you created the session with "createSession"?'
           }
         );
       });
@@ -872,7 +873,7 @@ describe('UseSessions', () => {
 
       beforeEach(() => ctx = createContext({ Authorization: `Bearer ${anonymousSessionID}`}));
 
-      postHookTestHook(async () => services.get(Store).readSession(authenticatedSessionID));
+      postHookTestHook(async () => readSession(services.get(Store), authenticatedSessionID));
 
     });
 
@@ -882,7 +883,7 @@ describe('UseSessions', () => {
 
         beforeEach(() => ctx = createContext());
 
-        postHookTestHook(async () => services.get(Store).readSession(authenticatedSessionID));
+        postHookTestHook(async () => readSession(services.get(Store), authenticatedSessionID));
 
       });
