+++ packages/core/src/sessions/http/utils/index.ts
@@ -1,6 +1,4 @@
-export { checkUserIdType } from './check-user-id-type';
 export { getCsrfTokenFromRequest } from './get-csrf-token-from-request';
-export { getSessionIDFromRequest, RequestValidationError } from './get-session-id-from-request';
 export { shouldVerifyCsrfToken } from './should-verify-csrf-token';
 export { removeSessionCookie } from './remove-session-cookie';
 export { setSessionCookie } from './set-session-cookie';
