+++ packages/core/src/sessions/index.ts
@@ -1,4 +1,6 @@
 export {
+  createSession,
+  readSession,
   Store,
   SessionStore,
   SessionAlreadyExists,
