+++ packages/jwt/src/http/utils/index.ts
@@ -1,5 +1,3 @@
-export { checkAndConvertUserIdType } from './check-and-convert-user-id-type';
 export { getCsrfTokenFromCookie } from './get-csrf-token-from-cookie';
 export { getCsrfTokenFromRequest } from './get-csrf-token-from-request';
-export { getJwtFromRequest, RequestValidationError } from './get-jwt-from-request';
 export { shouldVerifyCsrfToken } from './should-verify-csrf-token';
